requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
greenlet==3.0.3
idna==3.7
lxml==5.2.1
numpy==1.26.4
pandas==2.2.2
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
